# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/alumni/{id}": {
      "get": {
        "operationId": "AlumniController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Returns the alumni with the given id, enriched with location, graduations and roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alumni"
                }
              }
            }
          }
        },
        "summary": "Get alumni by id, enriched with location, graduations and roles",
        "tags": [
          "V1"
        ]
      }
    },
...
    "/api/analytics/industries": {
      "get": {
        "operationId": "IndustryAnalyticsController_getIndustryWithCounts",
        "parameters": [
          {
            "name": "alumniIds",
            "required": false,
            "in": "query",
            "description": "The alumni IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "The start date of the role",
            "schema": {
              "example": "2023-11-16",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "The end date of the role",
            "schema": {
              "example": "2025-04-20",
              "type": "string"
            }
          },
          {
            "name": "facultyIds",
            "required": false,
            "in": "query",
            "description": "The faculty IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "courseIds",
            "required": false,
            "in": "query",
            "description": "The course IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "graduationYears",
            "required": false,
            "in": "query",
            "description": "The graduation years to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyIds",
            "required": false,
            "in": "query",
            "description": "The company IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "industryIds",
            "required": false,
            "in": "query",
            "description": "The industry IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "roleCountryCodes",
            "required": false,
            "in": "query",
            "description": "The country codes where alumni exerced their roles",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "roleCityIds",
            "required": false,
            "in": "query",
            "description": "The city IDS where alumni exerced their roles",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyHQsCountryCodes",
            "required": false,
            "in": "query",
            "description": "The company HQs country codes to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyHQsCityIds",
            "required": false,
            "in": "query",
            "description": "The company HQs city ids to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "currentRolesOnly",
            "required": false,
            "in": "query",
            "description": "Filter for current roles only",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "onlyCompaniesWithSalaryData",
            "required": false,
            "in": "query",
            "description": "Include only companies with salary data",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "onlyInternational",
            "required": false,
            "in": "query",
            "description": "Whether to exclude roles in Portugal",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "excludeResearchAndHighEducation",
            "required": false,
            "in": "query",
            "description": "Exclude research and high education roles",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "alumniSearch",
            "required": false,
            "in": "query",
            "description": "Search query for companies",
            "schema": {
              "example": "search query",
              "type": "string"
            }
          },
          {
            "name": "companySize",
            "required": false,
            "in": "query",
            "description": "The company sizes to filter by",
            "schema": {
              "example": [
                "A",
                "B"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "A",
                  "B",
                  "C",
                  "D",
                  "E",
                  "F",
                  "G",
                  "H",
                  "I"
                ]
              }
            }
          },
          {
            "name": "companyType",
            "required": false,
            "in": "query",
            "description": "The company types to filter by",
            "schema": {
              "example": [
                "PUBLIC_COMPANY",
                "PRIVATELY_HELD"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "EDUCATIONAL",
                  "GOVERNMENT_AGENCY",
                  "NON_PROFIT",
                  "PARTNERSHIP",
                  "PRIVATELY_HELD",
                  "PUBLIC_COMPANY",
                  "SELF_EMPLOYED",
                  "SELF_OWNED"
                ]
              }
            }
          },
          {
            "name": "escoCodes",
            "required": false,
            "in": "query",
            "description": "The ESCO codes to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "escoClassificationLevel",
            "required": false,
            "in": "query",
            "description": "The ESCO classification level to filter by",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The number of results to return",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "The offset of the query",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Broad search query",
            "schema": {
              "example": "search query",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "How to sort the results",
            "schema": {
              "default": "count",
              "type": "string",
              "enum": [
                "count",
                "name",
                "year"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "The order of the results",
            "schema": {
              "default": "desc",
              "example": "asc",
              "type": "string"
            }
          },
          {
            "name": "includeTrend",
            "required": true,
            "in": "query",
            "description": "Whether to include the trend data",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The industries with the number of alumni working in them, and the number of companies in each industry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Returns the number of alumni working in companies grouped by industry.",
        "tags": [
          "V1"
        ]
      }
    },
...
    "/api/analytics/alumni": {
      "get": {
        "operationId": "AlumniAnalyticsController_getAlumniList",
        "parameters": [
          {
            "name": "alumniIds",
            "required": false,
            "in": "query",
            "description": "The alumni IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "The start date of the role",
            "schema": {
              "example": "2023-11-16",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "The end date of the role",
            "schema": {
              "example": "2025-04-20",
              "type": "string"
            }
          },
          {
            "name": "facultyIds",
            "required": false,
            "in": "query",
            "description": "The faculty IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "courseIds",
            "required": false,
            "in": "query",
            "description": "The course IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "graduationYears",
            "required": false,
            "in": "query",
            "description": "The graduation years to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyIds",
            "required": false,
            "in": "query",
            "description": "The company IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "industryIds",
            "required": false,
            "in": "query",
            "description": "The industry IDs to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "roleCountryCodes",
            "required": false,
            "in": "query",
            "description": "The country codes where alumni exerced their roles",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "roleCityIds",
            "required": false,
            "in": "query",
            "description": "The city IDS where alumni exerced their roles",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyHQsCountryCodes",
            "required": false,
            "in": "query",
            "description": "The company HQs country codes to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "companyHQsCityIds",
            "required": false,
            "in": "query",
            "description": "The company HQs city ids to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "currentRolesOnly",
            "required": false,
            "in": "query",
            "description": "Filter for current roles only",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "onlyCompaniesWithSalaryData",
            "required": false,
            "in": "query",
            "description": "Include only companies with salary data",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "onlyInternational",
            "required": false,
            "in": "query",
            "description": "Whether to exclude roles in Portugal",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "excludeResearchAndHighEducation",
            "required": false,
            "in": "query",
            "description": "Exclude research and high education roles",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "alumniSearch",
            "required": false,
            "in": "query",
            "description": "Search query for companies",
            "schema": {
              "example": "search query",
              "type": "string"
            }
          },
          {
            "name": "companySize",
            "required": false,
            "in": "query",
            "description": "The company sizes to filter by",
            "schema": {
              "example": [
                "A",
                "B"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "A",
                  "B",
                  "C",
                  "D",
                  "E",
                  "F",
                  "G",
                  "H",
                  "I"
                ]
              }
            }
          },
          {
            "name": "companyType",
            "required": false,
            "in": "query",
            "description": "The company types to filter by",
            "schema": {
              "example": [
                "PUBLIC_COMPANY",
                "PRIVATELY_HELD"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "EDUCATIONAL",
                  "GOVERNMENT_AGENCY",
                  "NON_PROFIT",
                  "PARTNERSHIP",
                  "PRIVATELY_HELD",
                  "PUBLIC_COMPANY",
                  "SELF_EMPLOYED",
                  "SELF_OWNED"
                ]
              }
            }
          },
          {
            "name": "escoCodes",
            "required": false,
            "in": "query",
            "description": "The ESCO codes to filter by",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "escoClassificationLevel",
            "required": false,
            "in": "query",
            "description": "The ESCO classification level to filter by",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The number of results to return",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "The offset of the query",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Broad search query",
            "schema": {
              "example": "search query",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "How to sort the results",
            "schema": {
              "default": "count",
              "type": "string",
              "enum": [
                "count",
                "name",
                "year"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "The order of the results",
            "schema": {
              "default": "desc",
              "example": "asc",
              "type": "string"
            }
          },
          {
            "name": "includeTrend",
            "required": true,
            "in": "query",
            "description": "Whether to include the trend data",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alumni list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlumniListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all alumni list",
        "tags": [
          "V1"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Company": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the company"
          },
          "name": {
            "type": "string",
            "description": "The name of the company"
          },
          "linkedinUrl": {
            "type": "object",
            "description": "The linkedin url of the company"
          },
          "roles": {
            "description": "The roles of the company",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "roles"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the role"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "description": "The start date of the role"
          },
          "endDate": {
            "type": "object",
            "description": "The end date of the role"
          },
          "seniorityLevel": {
            "type": "string",
            "description": "The seniority level of the role",
            "enum": [
              "INTERN",
              "ENTRY_LEVEL",
              "ASSOCIATE",
              "MID_SENIOR_LEVEL",
              "DIRECTOR",
              "EXECUTIVE",
              "C_LEVEL"
            ]
          },
          "Company": {
            "description": "The Company",
            "allOf": [
              {
                "$ref": "#/components/schemas/Company"
              }
            ]
          },
          "JobClassification": {
            "type": "object",
            "description": "The job classification"
          },
          "Location": {
            "type": "object",
            "description": "The location of the role"
          }
        },
        "required": [
          "id",
          "startDate",
          "seniorityLevel",
          "Company"
        ]
      },
      "ReviewCompany": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the review"
          },
          "description": {
            "type": "object",
            "description": "The review"
          },
          "rating": {
            "type": "number",
            "description": "The rating of the review"
          },
          "upvotes": {
            "description": "upvotes fo the review",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "downvotes": {
            "description": "downvvotes of the review",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "Location": {
            "type": "object",
            "description": "The location of the review"
          },
          "Company": {
            "description": "The Company",
            "allOf": [
              {
                "$ref": "#/components/schemas/Company"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date and time when the review was created"
          }
        },
        "required": [
          "id",
          "description",
          "rating",
          "upvotes",
          "downvotes",
          "Company"
        ]
      },
      "ReviewLocation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the review"
          },
          "description": {
            "type": "object",
            "description": "The review"
          },
          "rating": {
            "type": "number",
            "description": "The rating of the review"
          },
          "upvotes": {
            "description": "upvotes fo the review",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "downvotes": {
            "description": "downvvotes of the review",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "Location": {
            "type": "object",
            "description": "The location of the review"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date and time when the review was created"
          }
        },
        "required": [
          "id",
          "description",
          "rating",
          "upvotes",
          "downvotes"
        ]
      },
      "LocationGeo": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "description": "The city of the location"
          },
          "country": {
            "type": "string",
            "description": "The country of the location"
          },
          "countryCode": {
            "type": "string",
            "description": "The country code of the location"
          },
          "latitude": {
            "type": "number",
            "description": "The latitude of the location"
          },
          "longitude": {
            "type": "number",
            "description": "The longitude of the location"
          }
        }
      },
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the course"
          },
          "name": {
            "type": "string",
            "description": "The name of the course"
          },
          "acronym": {
            "type": "string",
            "description": "The acronym of the course"
          }
        },
        "required": [
          "id",
          "name",
          "acronym"
        ]
      },
      "Graduation": {
        "type": "object",
        "properties": {
          "conclusionYear": {
            "type": "object",
            "description": "The conclusion year of the graduation"
          },
          "Course": {
            "description": "The course of the graduation",
            "allOf": [
              {
                "$ref": "#/components/schemas/Course"
              }
            ]
          }
        },
        "required": [
          "conclusionYear",
          "Course"
        ]
      },
      "Alumni": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the alumni"
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the alumni"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the alumni"
          },
          "fullName": {
            "type": "string",
            "description": "The full name of the alumni"
          },
          "linkedinUrl": {
            "type": "string",
            "description": "The linkedin url of the alumni"
          },
          "profilePictureUrl": {
            "type": "string",
            "description": "The profile picture of the alumni"
          },
          "source": {
            "type": "string",
            "description": "The source of the alumni",
            "enum": [
              "FORM_SUBMISSION",
              "ADMIN_IMPORT",
              "HONORARY_MEMBER"
            ]
          },
          "Roles": {
            "description": "The roles of the alumni",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "ReviewsCompany": {
            "description": "The reviews of the alumni",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewCompany"
            }
          },
          "ReviewsLocation": {
            "description": "",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewLocation"
            }
          },
          "Location": {
            "description": "The current location of the alumni",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocationGeo"
              }
            ]
          },
          "Graduations": {
            "description": "The graduation status(es) of the alumni",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Graduation"
            }
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName"
        ]
      },
...
      "DataPointDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "The label of the data point",
            "example": "2021-01-01"
          },
          "value": {
            "type": "number",
            "description": "The value of the data point",
            "example": 100
          }
        },
        "required": [
          "label",
          "value"
        ]
      },
...
      "IndustryListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the industry"
          },
          "name": {
            "type": "string",
            "description": "The name of the industry",
            "example": "Financial Services"
          },
          "count": {
            "type": "number",
            "description": "The number of alumni working in the industry",
            "example": 100
          },
          "trend": {
            "description": "The alumni trend of the industry",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataPointDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "count",
          "trend"
        ]
      },
      "IndustryListResponseDto": {
        "type": "object",
        "properties": {
          "industries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndustryListItemDto"
            }
          },
          "count": {
            "type": "number",
            "description": "The total number of industries in the database"
          },
          "filteredCount": {
            "type": "number",
            "description": "The total number of industries in the database before applying the filters"
          }
        },
        "required": [
          "industries",
          "count",
          "filteredCount"
        ]
      },
...
      "AlumniListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The alumni ID"
          },
          "fullName": {
            "type": "string",
            "description": "The full name of the alumni"
          },
          "linkedinUrl": {
            "type": "string",
            "description": "The linkedin url of the alumni"
          },
          "profilePictureUrl": {
            "type": "string",
            "description": "The profile picture of the alumni"
          },
          "currentRoleLocation": {
            "description": "The location of the alumni`s current role",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocationAnalyticsEntity"
              }
            ]
          }
        },
        "required": [
          "id",
          "fullName"
        ]
      },
      "AlumniListResponseDto": {
        "type": "object",
        "properties": {
          "alumni": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlumniListItemDto"
            }
          },
          "count": {
            "type": "number",
            "description": "The total number of alumni in the database before applying the filters"
          },
          "filteredCount": {
            "type": "number",
            "description": "The total number of alumni in the database after applying the filters"
          }
        },
        "required": [
          "alumni",
          "count",
          "filteredCount"
        ]
      },
...
    }
  }
}